+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Option for displaying a logo.
 - Convars for changing the screen colors & CSS background.
 - Convar option for setting a custom font family.
+- Convar option for setting the background brightness.
 
 ### Changed